add asset tag model
An asset tag is simply a code which is marked as either 'active' or
'inactive'. Update the Device model to have some number of asset tags
associated with it. The rationale for having multiple asset tags
associated with one device is that we may change the asset tagging
system as time develops meaning that devices may (temporarily) have
multiple tags. The rationale for not mandating that all devices be
tagged is that being able to record a) a new device and b) the fact that
it is not yet asset tagged is useful in and of itself.