add Owner model to database
An owner is an entity which owns a device in some sense. A device may
have more than one owner. Initially an owner is a pair or strings giving
a name and email address to contact.

Some basic validation of the email address is performed: it is
normalised to lower-case, each owner must have a unique address and the
address must have a '@' sign somewhere which is not the first or last
character. Trying to do any more validation of email addresses without
actually sending it is doomed to failure :).