create_statuses_and_reference_urls: insert values singularly
The multiple value insert syntax was technically wrong for sqlite but OK
(I think) for Postgres. To avoid confusion, insert values singularly.

It would be better, I think, for there to be a Status model which hid
the necessity to put SQL directly into the migration to avoid this sort
of incompatibility.