Status: add severity field
Statuses should have an explicit severity recorded in the database
rather than one hard-coded into the view. It may be that many statuses
warrant being called an 'error' or a 'warning' and it is best if that
information is in the database somewhere.

This commit includes a migration which automatically converts existing
statuses in the database over to having a severity based on the old
rules.