only use sendgrid in production
I don't have sendgrid set up on my machine and travis-ci certainly
doesn't. Since the devise framework sends email as part of the sign up
test, having this configured was causing the test suite to fail when
mergind pull request #14.

Wrapping the configuration in `if Rails.env production?` seems a little
ugly but apparently short of pasting it into
config/environments/production.rb, this is the only way.